requirements.txt
flask
torch==2.1.1
requests
python-dotenv
supabase
openai==0.27.7
transformers
autoawq
gunicorn